20
public int addScore adds new score to table if it is lower than any score currently held. The name field in the table is set to "ENTER NAME", and the score field is set to the player's score.
The scores are in order, and the previous scores are shifted downwards if they are higher than the new score. The last score is "popped off".